adds commentary and weatherTest file
added comments in the existing files to explain how they are working so that I
don't have to re-learn things from day-to-day (I'm new here. give me some
slack). Also renamed the "disp" to "oled" in mathisPi.py.

created the mathisWeatherTest.py to see how hard it would be to get the
weather data through curl (or similar) within a python script. Turns out the
answer is, "pretty easy if you have access to google..."